Refactor : 이전 '멘토 상세 조회' 관련 코드 제거
