+++ src/main/java/sideproject/coffeechat/domain/baton/entity/BatonStatus.java
@@ -1,18 +1,11 @@
 package sideproject.coffeechat.domain.baton.entity;
 
-import lombok.Getter;
-import lombok.RequiredArgsConstructor;
-
-@Getter
-@RequiredArgsConstructor
 public enum BatonStatus {
 
-    REQUESTED("requestedCount"),
-    ACCEPTED("acceptedCount"),
-    REJECTED("rejectedCount"),
-    COMPLETED("completedCount"),
+    REQUESTED,
+    ACCEPTED,
+    REJECTED,
+    COMPLETED,
     ;
-
-    private final String countKey;
 
 }
