+++ src/main/java/sideproject/coffeechat/domain/baton/service/BatonViewService.java
@@ -1,20 +1,12 @@
 package sideproject.coffeechat.domain.baton.service;
 
-import java.util.Map;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
-import sideproject.coffeechat.domain.baton.repository.BatonMapper;
 
 @Service
 @RequiredArgsConstructor
 @Transactional(readOnly = true)
 public class BatonViewService {
-
-    private final BatonMapper batonMapper;
-
-    public Map<String, Long> getBatonStatusCounts(Long mentorId) {
-        return batonMapper.getBatonStatusCounts(mentorId);
-    }
 
 }
