+++ src/main/java/sideproject/coffeechat/domain/mentor/dto/mapperdto/CompactMentorDTO.java
@@ -1,4 +1,4 @@
-package sideproject.coffeechat.domain.mentor.dto;
+package sideproject.coffeechat.domain.mentor.dto.mapperdto;
 
 import java.time.LocalDateTime;
 import lombok.Getter;
